requirements.txt
discord.py>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
